Create Mega_CAN_SendTemperature.ino
Code for an Arduino Mega set up with the AHT20 i2c temperature sensors. It broadcasts the data from each sensor over the CAN bus every [interval] milliseconds. The interval is currently set to 3000ms. Minimum interval in this configuration seems to be 2 seconds.

Additionally, this is configured to respond to the ping message from the CAN Master Controller.