Create Nano_CAN_Master_Controller.ino
Interfaces between the Python program running on the Raspberry Pi control device and the CAN bus with the Arduinos.  This program sends a ping on startup to determine which devices are connected, and has the capability to, when directed by the Python program, request data from all the Arduinos on the CAN bus.